Added "entity_health" script.
Created "Scripts" folder. We should put all new custom scripts in here
to stay organized.

Created "entity_health.cs". This script holds a health value that can be
securely adjusted and referenced (for the use of other scripts).
Intended to be "portable", i.e. just attach this script as a component
to a gameobject and it will now have a health value!